pyproject.toml
[project]
name = "langchain-foundations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.6",
    "python-dotenv>=1.2.1",
    "langchain-model-profiles>=0.0.4",
    "langchain-tavily>=0.2.13",
    "tavily-python>=0.7.13",
    "tavily>=1.1.0",
    "pypdf>=6.2.0",
    "mcp>=1.21.1",
    "langchain-mcp-adapters>=0.1.13",
    "langgraph>=1.0.7",
    "langgraph-cli>=0.4.12",
    "langgraph-api>=0.7.9",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-google-genai>=4.2.0",
    "langsmith>=0.6.4",
    "notebook>=7.5.2",
    "ipywidgets>=8.1.8",
    "sounddevice>=0.5.5",
    "scipy>=1.17.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
openai = [
    "langchain-openai>=1.1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
